adding the file from Yan's computer
at the moment, the code will run and a graph of the final concentrations of all elems will be made. however, it seems like the if statement parts are not working as intended, since we only get concentrations of 0 or maxConc.

am currently using placeholder values, no need to worry if the variables do not match what was written in comments.